AtomsView

AtomsView is a package of tools for visualising AtomsBase-compatible structures. The functionality is still limited and interfaces unstable. Expect them to change at any time.

AtomsView.visualize_structure

AtomsView.visualize_structure — Method

Produce some text/html code which allows to interactively view the system. Currently uses Bio3DView and some kwargs can be used to customise the representation. Note that the backend as well as interface is likely going to change in the future.